bootcamp/week-11-testing_seguridad_docker/2-practicas/practica-04-seguridad-hardening/starter/pyproject.toml
[project]
name = "practica-04-seguridad-hardening"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "mcp==1.9.0",
    "aiosqlite==0.20.0",
    "httpx==0.27.2",
    "python-dotenv==1.0.1",
]
